pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stratum"
version = "0.1.0"
description = "Domain-specific RAG engine with hybrid retrieval, cross-encoder re-ranking, and citation-grounded generation"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  # Vector stores
  "chromadb>=0.5.0",             # default backend — zero infrastructure
  "weaviate-client>=4.5.0",      # production backend — opt-in via STRATUM_STORE_BACKEND=weaviate

  # Embeddings — OpenAI is the default (no GPU, no large download)
  "anthropic>=0.30.0",
  "openai>=1.30.0",
  "numpy>=1.26.0",

  # Retrieval
  "rank-bm25>=0.2.2",
  "sentence-transformers>=2.7.0", # cross-encoder reranker (always required, ~180MB)

  # Document loading
  "pypdf>=4.0.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",

  # Config / logging
  "pydantic-settings>=2.3.0",
  "structlog>=24.1.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
# Heavy local embeddings — opt-in only (requires GPU recommended, ~2GB download)
local-embed = [
  "torch>=2.2.0",
]
# RAGAS eval — separate optional group (LLM-evaluated, costs money per run)
eval = [
  "ragas==0.1.21",         # pinned — RAGAS has broken API on minor bumps historically
  "datasets>=2.19.0",
  "pandas>=2.2.0",
]
api = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.10.0",
  "ruff>=0.4.0",
  "pre-commit>=3.7.0",
  "stratum[eval]",
  "stratum[api]",
]

[project.scripts]
stratum-ingest = "rag.scripts.ingest:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/rag --cov-report=term-missing --no-cov-on-fail"
markers = [
  "integration: requires live infrastructure (Chroma runs in CI; Weaviate requires Docker)",
  "e2e: requires STRATUM_ANTHROPIC_API_KEY and eval dependencies",
  "slow: marks tests as slow (deselect with -m 'not slow')",
]

[tool.coverage.report]
fail_under = 80
